[update] the component stack to get it's own module under components & update references to it within lambdarunnable.
