+++ lambda/src/components/mod.rs
@@ -1,2 +1,5 @@
 mod renderer;
 pub use renderer::Renderer;
+
+mod component_stack;
+pub use component_stack::ComponentStack;
